src/MQTT/requirements.txt
Flask
paho-mqtt
confluent_kafka
python-dotenv
requests